1、修改为使用requests 模块下载，待测试

+++ handle.py
@@ -9,7 +9,7 @@ from utils.util_logfile import nlogger, flogger, traceback
 from utils.util_xlsx import HandleXLSX
 from utils.util_requester import BeautifulSoup, requester, RequestException
 from utils.util_re import is_url
-from utils.util_download import get_file_path, download_video, check_file_exist, check_temp_file_exists, download_file
+from utils.util_download import get_file_path, download_video, check_file_exist, download_file, Download
 from datetime import datetime
 from settings import BASE_URL, TASK_WAITING_TIME, FORCE_DOWNLOAD, MAX_WORKERS, DL_ROOT_PATH
 from row_object import RowStatus
@@ -39,9 +39,10 @@ class WriteResultError(Exception):
     pass
 
 
-def exec_func(file_name, sheet_name=None, start_point=None, end_point=None, **kwargs):
+def exec_func(check_file, file_name, sheet_name=None, start_point=None, end_point=None, **kwargs):
     """
     Executive Function
+    :param check_file: if check_file is True,then only check if download file exists. default False
     :param file_name: Excel file name
     :param sheet_name: sheet name, default active sheet
     :param start_point: start row number, minimum is 2 ( row 1 is column name)
@@ -51,8 +52,8 @@ def exec_func(file_name, sheet_name=None, start_point=None, end_point=None, **kw
     """
     try:
         _file_name = check_file_name(file_name, **kwargs)
-        _source_xls, _row_object_iterator = get_row_object_iterator(_file_name, sheet_name, start_point, end_point,
-                                                                    **kwargs)
+        _source_xls, _row_object_iterator = get_row_object_iterator(check_file, _file_name, sheet_name, start_point,
+                                                                    end_point, **kwargs)
         _download_result = download_video_thread(_row_object_iterator, **kwargs)
         write_result_to_xls(_source_xls, _download_result)
     except (GetURLError, GetRowIterError, GetStorageError, DownloadVideoError, ThreadTaskError, WriteResultError) as e:
@@ -79,9 +80,10 @@ def check_file_name(file_name, **kwargs):
     return _file_name
 
 
-def get_row_object_iterator(file_name, sheet_name=None, start_point=None, end_point=None, **kwargs):
+def get_row_object_iterator(check_file, file_name, sheet_name=None, start_point=None, end_point=None, **kwargs):
     """
     get iterator of row object
+    :param check_file: True or False
     :param file_name:
     :param sheet_name:
     :param start_point:
@@ -91,7 +93,8 @@ def get_row_object_iterator(file_name, sheet_name=None, start_point=None, end_po
     """
     try:
         source_xls = HandleXLSX(file_name, sheet_name)
-        row_object_iterator = source_xls.generate_row_object_iterator(sheet_name, start_point, end_point, **kwargs)
+        row_object_iterator = source_xls.generate_row_object_iterator(check_file, sheet_name, start_point, end_point,
+                                                                      **kwargs)
         return source_xls, row_object_iterator
     except Exception as e:
         nlogger.error('{fn} error: {e}'.format(fn='get_row_object_iterator', e=traceback.format_exc()))
@@ -202,7 +205,8 @@ def check_row_object(row_object, **kwargs):
                                                   dict), "Parameter column_value is dict and not be empty"
     assert row_object.position and isinstance(row_object.position,
                                               int), "Parameter position is int and not be less than 1"
-    assert row_object.status and RowStatus(row_object.status).name == 'INITIAL', "Parameter status is invalid"
+    assert row_object.status and RowStatus(row_object.status).name in ['INITIAL',
+                                                                       'CHECK'], "Parameter status is invalid"
 
 
 def get_video_url(pre_row_object, **kwargs):
@@ -214,7 +218,7 @@ def get_video_url(pre_row_object, **kwargs):
     """
     try:
         # File already exists, skip getting URL
-        if RowStatus(pre_row_object.status).name == 'SKIP':
+        if RowStatus(pre_row_object.status).name in ['SKIP', 'EXISTENCE', 'NONEXISTENCE']:
             pre_row_object.temporary_url = None
             return pre_row_object
         else:
@@ -265,14 +269,21 @@ def get_storage_absolute_path(row_object, root_path, force_download, **kwargs):
         row_object.storage_absolute_path = _storage_absolute_path
 
         _exist = check_file_exist(_storage_absolute_path, **kwargs)
-        if _exist is False:
-            row_object.status = RowStatus.DOWNLOAD.value
-        else:
-            if force_download is True:
-                os.remove(_storage_absolute_path)
-                row_object.status = RowStatus.FORCE.value
+
+        if RowStatus(row_object.status).name == 'CHECK':
+            if _exist is False:
+                row_object.status = RowStatus.NONEXISTENCE.value
             else:
-                row_object.status = RowStatus.SKIP.value
+                row_object.status = RowStatus.EXISTENCE.value
+        else:
+            if _exist is False:
+                row_object.status = RowStatus.DOWNLOAD.value
+            else:
+                if force_download is True:
+                    os.remove(_storage_absolute_path)
+                    row_object.status = RowStatus.FORCE.value
+                else:
+                    row_object.status = RowStatus.SKIP.value
 
         return row_object
     except Exception as e:
@@ -294,8 +305,13 @@ def download_video_from_url(mid_row_object, **kwargs):
 
         if RowStatus(mid_row_object.status).name == 'SKIP':
             mid_row_object.column_value['result'] = f'Already exists: {absolute_path_file}'
-            nlogger.info(
-                'Download skip: {f}, its storage path is {p}.'.format(f=file_name, p=absolute_path_file))
+            nlogger.info('Download skip: {f}, its storage path is {p}.'.format(f=file_name, p=absolute_path_file))
+        elif RowStatus(mid_row_object.status).name == 'EXISTENCE':
+            mid_row_object.column_value['result'] = f'EXISTENCE: {absolute_path_file}'
+            nlogger.info('Download ignore: {f}, it exists.'.format(f=file_name))
+        elif RowStatus(mid_row_object.status).name == 'NONEXISTENCE':
+            mid_row_object.column_value['result'] = f'NONEXISTENCE: {absolute_path_file}'
+            nlogger.info('Download ignore: {f}, it does not exist.'.format(f=file_name))
         else:
             dl_file_name = download_video(url, absolute_path_file, file_name, **kwargs)
             if dl_file_name:
@@ -313,6 +329,36 @@ def download_video_from_url(mid_row_object, **kwargs):
         raise DownloadVideoError("{fn} error: {e}".format(fn='download_video_from_url', e=repr(e)))
 
 
+dl = Download()
+
+
+def get_file_size_megabyte(dl_file_name):
+    return round(dl.get_file_size(dl_file_name) / float(1024 * 1024), 4)
+
+
+def set_row_object_status(mid_row_object, file_name, dl_file_name=None, **kwargs):
+    if dl_file_name:
+        if RowStatus(mid_row_object.status).name == 'SKIP':
+            mid_row_object.column_value['result'] = f'Already exists: {dl_file_name}'
+            mid_row_object.column_value['size(MB)'] = f'{get_file_size_megabyte(dl_file_name)}'
+            mid_row_object.column_value['type'] = ''
+        elif RowStatus(mid_row_object.status).name == 'FORCE':
+            mid_row_object.column_value['result'] = f'Download again: {dl_file_name}'
+            mid_row_object.column_value['size(MB)'] = f'{get_file_size_megabyte(dl_file_name)}'
+            mid_row_object.column_value['type'] = ''
+        else:
+            mid_row_object.column_value['result'] = f'Download successful: {dl_file_name}'
+            mid_row_object.column_value['size(MB)'] = f'{get_file_size_megabyte(dl_file_name)}'
+            mid_row_object.column_value['type'] = ''
+        mid_row_object.status = RowStatus.SUCCESS.value
+    else:
+        mid_row_object.column_value['result'] = f"Download failed: {file_name}"
+        # mid_row_object.column_value['size(MB)'] = ''
+        # mid_row_object.column_value['type'] = ''
+        mid_row_object.status = RowStatus.FAILURE.value
+    return mid_row_object
+
+
 def download_file_from_url(mid_row_object, **kwargs):
     """
     Download file from temporary URL
@@ -322,28 +368,44 @@ def download_file_from_url(mid_row_object, **kwargs):
     """
     try:
         url = mid_row_object.temporary_url
-        file_name = "{f}".format(f=mid_row_object.column_value.get('vid'))
-        absolute_path_file = mid_row_object.storage_absolute_path
+        file_name = f"{mid_row_object.column_value.get('vid')}"
+        storage_absolute_path = mid_row_object.storage_absolute_path
 
-        if RowStatus(mid_row_object.status).name == 'SKIP':
-            mid_row_object.column_value['result'] = f'Already exists: {absolute_path_file}'
-            nlogger.info(
-                'Download skip: {f}, its storage path is {p}.'.format(f=file_name, p=absolute_path_file))
+        if RowStatus(mid_row_object.status).name == 'EXISTENCE':
+            mid_row_object.column_value['result'] = f'EXISTENCE: {storage_absolute_path}'
+            nlogger.info('Download ignore: {f}, its storage path is {p}.'.format(f=file_name, p=storage_absolute_path))
+        elif RowStatus(mid_row_object.status).name == 'NONEXISTENCE':
+            mid_row_object.column_value['result'] = f'NONEXISTENCE: {storage_absolute_path}'
+            nlogger.info('Download ignore: {f}, it does not exist.'.format(f=file_name))
+        elif RowStatus(mid_row_object.status).name == 'SKIP':
+            mid_row_object = set_row_object_status(mid_row_object, file_name, dl_file_name=storage_absolute_path)
         else:
-            dl_file_name = download_file(url, absolute_path_file, file_name, chunk_size=4096 * 1024, retry=3, **kwargs)
-            if dl_file_name:
-                if RowStatus(mid_row_object.status).name == 'FORCE':
-                    mid_row_object.column_value['result'] = f'Download again: {dl_file_name}'
-                else:
-                    mid_row_object.column_value['result'] = f'Download successful: {dl_file_name}'
-                mid_row_object.status = RowStatus.SUCCESS.value
-            else:
-                mid_row_object.column_value['result'] = f"Download failed: {file_name}"
-                mid_row_object.status = RowStatus.FAILURE.value
+            kwargs['download_url'] = url
+            kwargs['storage_file_name'] = storage_absolute_path
+            kwargs['chunk_size'] = 4096 * 1024
+            kwargs['retry'] = 3
+            kwargs['timeout'] = (10, 120)
+
+            try:
+                dl_file_name = dl.download(**kwargs)
+            except Exception as e:
+                nlogger.error("{fn} dl.download error: {e}".format(fn='download_file_from_url', e=repr(e)))
+                dl_file_name = None
+
+            mid_row_object = set_row_object_status(mid_row_object, file_name, dl_file_name=dl_file_name)
+            # if dl_file_name:
+            #     if RowStatus(mid_row_object.status).name == 'FORCE':
+            #         mid_row_object.column_value['result'] = f'Download again: {dl_file_name}'
+            #     else:
+            #         mid_row_object.column_value['result'] = f'Download successful: {dl_file_name}'
+            #     mid_row_object.status = RowStatus.SUCCESS.value
+            # else:
+            #     mid_row_object.column_value['result'] = f"Download failed: {file_name}"
+            #     mid_row_object.status = RowStatus.FAILURE.value
         return mid_row_object
     except Exception as e:
-        nlogger.error("{fn} error: {e}".format(fn='download_video_from_url', e=traceback.format_exc()))
-        raise DownloadVideoError("{fn} error: {e}".format(fn='download_video_from_url', e=repr(e)))
+        nlogger.error("{fn} error: {e}".format(fn='download_file_from_url', e=traceback.format_exc()))
+        raise DownloadVideoError("{fn} error: {e}".format(fn='download_file_from_url', e=repr(e)))
 
 
 def write_result_to_xls(source_xls, download_result):
@@ -358,8 +420,26 @@ def write_result_to_xls(source_xls, download_result):
         for row_object in download_result:
             x = row_object.position
             y = column_name_list.index('result') + 1
+            column_number = len(column_name_list)
             values = (x, y, row_object.column_value.get('result', 'unknown'))
             source_xls.write_sheet_rows_value(sheet_name=row_object.sheet_name, values=values)
+
+            if row_object.column_value.get('size(MB)'):
+                if 'size(MB)' in column_name_list:
+                    y = column_name_list.index('size(MB)') + 1
+                else:
+                    y = column_number + 1
+                    column_number += 1
+                values = (x, y, row_object.column_value.get('size(MB)'))
+                source_xls.write_sheet_rows_value(sheet_name=row_object.sheet_name, values=values)
+
+            if row_object.column_value.get('type'):
+                if 'type' in column_name_list:
+                    y = column_name_list.index('type') + 1
+                else:
+                    y = column_number + 1
+                values = (x, y, row_object.column_value.get('type'))
+                source_xls.write_sheet_rows_value(sheet_name=row_object.sheet_name, values=values)
 
         _dl_file_name = "dl_{d}.xlsx".format(d=datetime.now().strftime('%Y%m%d-%H:%M:%S'))
         source_xls.save(_dl_file_name)
@@ -398,23 +478,26 @@ def main(argv):
     :param argv: command parameters
     :return:
     """
+    _check_file = False
     _file_name = None
     _sheet_name = None
     _start_point = None
     _end_point = None
 
     try:
-        opts, args = getopt.getopt(argv, "hf:t:s:e:", ["help", "file", "sheet", "start", "end"])  # 短选项模式
+        opts, args = getopt.getopt(argv, "hcf:t:s:e:", ["help", "file", "sheet", "start", "end"])  # 短选项模式
     except getopt.GetoptError:
         print("Usage 1: python3.9 handle.py -h  -f <source excel>  -t <excel sheet>  -s <start row index>  "
-              "-e <end row index> \nUsage 2: python3.9 handle.py --help  --file <source excel>  --sheet <excel sheet>  "
-              "--start <start row index>  --end <end row index>")
+              "-e <end row index> \nUsage 2: python3.9 handle.py --help --check --file <source excel>  "
+              "--sheet <excel sheet> --start <start row index>  --end <end row index>")
         sys.exit(2)  # 2 Incorrect Usage
 
     for opt, arg in opts:
         if opt in ('-h', '--help'):
             usage()
             sys.exit(0)
+        elif opt in ('-c', '--check'):
+            _check_file = True
         elif opt in ('-f', '--file'):
             _file_name = str(arg).strip()
         elif opt in ('-t', '--sheet'):
@@ -436,7 +519,8 @@ def main(argv):
         print("Invalid parameter, -e --end must be followed by integer. \nTry '-h --help' for more information.")
         sys.exit(2)
 
-    params = dict(file_name=_file_name, sheet_name=_sheet_name, start_point=_start_point, end_point=_end_point)
+    params = dict(check_file=_check_file, file_name=_file_name, sheet_name=_sheet_name, start_point=_start_point,
+                  end_point=_end_point)
 
     exec_func(**params)
